Added add and edit review templates

+++ reviews/views.py
@@ -2,22 +2,20 @@ from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 
-from profiles.models import UserProfile
 from products.models import Product
 from .models import Review
+from .forms import ProductReviewForm
 
 
 @login_required
 def add_review(request, product_id):
     """ To add a product review """
-
     product = get_object_or_404(Product, pk=product_id)
     if request.user.is_authenticated:
-
         if request.method == 'POST':
             form = ProductReviewForm(request.POST)
             if form.is_valid():
-                review = form.save()
+                review = form.save(commit=False)
                 review.user = request.user
                 review.product = product
                 review.save()
@@ -25,23 +23,26 @@ def add_review(request, product_id):
                 return redirect(reverse('product_detail', args=[product.id]))
             else:
                 messages.error(request, 'Failed to add review. Please ensure the form is valid.')
-    context = {
-        'form': form,
-        'product': product
-    }
+        else:
+            form = ProductReviewForm()
 
-    return render(request, context)
+        template = 'reviews/add_product_review.html'
+        context = {
+            'form': form,
+            'product': product
+        }
+
+        return render(request, template, context)
 
 
 @login_required
-def edit_review(request, product_id, review_id ):
+def edit_review(request, product_id, review_id):
     """ To edit a product review """
 
-    review = get_object_or_404(Review, pk=product_id)
+    review = get_object_or_404(Review, pk=review_id)
     product = review.product
 
     if request.user.is_authenticated:
-
         if request.method == 'POST':
             form = ProductReviewForm(request.POST)
             if form.is_valid():
@@ -53,26 +54,28 @@ def edit_review(request, product_id, review_id ):
                 return redirect(reverse('product_detail', args=[product.id]))
             else:
                 messages.error(request, 'Failed to update the review. Please ensure the form is valid.')
-    else:
-        form = ProductReviewForm(instance=review)
+        else:
+            form = ProductReviewForm()
 
-    context = {
-        'review': review,
-        'form': form,
-        'product': product,
-    }
+        template = 'reviews/add_product_review.html'
+        context = {
+            'review': review,
+            'form': form,
+            'product': product,
+        }
 
-    return render(request, 'reviews/edit_review.html', context)
+        return render(request, template, context)
 
 
 @login_required
-def delete_review(request, product_id, review_id ):
+def delete_review(request, product_id, review_id):
     """ To delete a product review """
+    if not request.user:
+        messages.error(request, "You can't edit this review")
+        return redirect(reverse('home'))
 
-    review = get_object_or_404(Review, pk=product_id)
+    review = get_object_or_404(Review, pk=review_id)
     product = review.product
-
     review.delete()
-    
     messages.success(request, f'Deleted review for {product.name}')
     return redirect(reverse('product_detail', args=[product.id]))
